+++ src/app/(more)/_layout.tsx
@@ -12,6 +12,7 @@ export default function MoreLayout(): ReactElement {
       <Stack.Screen name="calculation" options={{ title: strings.calculation.title }} />
       <Stack.Screen name="hijri" options={{ title: strings.hijri.title }} />
       <Stack.Screen name="tracking" options={{ title: strings.tracking.title }} />
+      <Stack.Screen name="moon-sighting" options={{ title: strings.moonSighting.title }} />
     </Stack>
   )
 }